mizumoto_project/questionnare/requirements.txt
matplotlib==3.8.2
gspread==5.12.4
google-auth==2.27.0 